<==== Mercurial data ====>
Branch: async_ssl

+++ client/src/utils/app_server_file_cache.cpp
@@ -2,36 +2,73 @@
 
 #include <QtCore/QDir>
 #include <QtCore/QFileInfo>
-#include <utils/common/uuid.h>
 #include <QtCore/QTimer>
 #include <QtCore/QStandardPaths>
 
-#include <utils/common/util.h> /* For removeDir. */
-
 #include <api/app_server_connection.h>
 
+#include <common/common_module.h>
+#include <client/client_message_processor.h>
 
-QnAppServerFileCache::QnAppServerFileCache(QString folderName, QObject *parent) :
-    QObject(parent),
-    m_fileListHandle(0),
-    m_folderName(folderName)
-{
-    connect(this, SIGNAL(delayedFileDownloaded(QString,bool)), this, SIGNAL(fileDownloaded(QString,bool)), Qt::QueuedConnection);
-    connect(this, SIGNAL(delayedFileUploaded(QString,bool)), this, SIGNAL(fileUploaded(QString,bool)), Qt::QueuedConnection);
-    connect(this, SIGNAL(delayedFileDeleted(QString,bool)), this, SIGNAL(fileDeleted(QString,bool)), Qt::QueuedConnection);
+#include <utils/common/util.h> /* For removeDir. */
+#include <utils/common/uuid.h>
+#include <utils/common/string.h>
+
+namespace {
+    /** Maximum allowed file size is 15 Mb, hard limit of Jaguar or S2 cameras. */
+    const qint64 maximumFileSize = 15*1024*1024;
 }
 
-QnAppServerFileCache::~QnAppServerFileCache(){}
+QnAppServerFileCache::QnAppServerFileCache(const QString &folderName, QObject *parent) :
+    QObject(parent),
+    m_folderName(folderName)
+{
+    connect(this, &QnAppServerFileCache::delayedFileDownloaded,     this,   [this](const QString &filename, OperationResult status) {
+        if (isConnectedToServer())
+            emit fileDownloaded(filename, status);
+        else
+            emit fileDownloaded(filename, OperationResult::disconnected);
+    }, Qt::QueuedConnection);
+
+    connect(this, &QnAppServerFileCache::delayedFileUploaded,       this,   [this](const QString &filename, OperationResult status) {
+        if (isConnectedToServer())
+            emit fileUploaded(filename, status);
+        else
+            emit fileUploaded(filename, OperationResult::disconnected);
+    }, Qt::QueuedConnection);
+
+    connect(this, &QnAppServerFileCache::delayedFileDeleted,        this,   [this](const QString &filename, OperationResult status) {
+        if (isConnectedToServer())
+            emit fileDeleted(filename, status);
+        else
+            emit fileDeleted(filename, OperationResult::disconnected);
+    }, Qt::QueuedConnection);
+
+    connect(this, &QnAppServerFileCache::delayedFileListReceived,   this,   [this](const QStringList &files, OperationResult status) {
+        if (isConnectedToServer())
+            emit fileListReceived(files, status);
+        else
+            emit fileListReceived(files, OperationResult::disconnected);
+    }, Qt::QueuedConnection);
+}
+
+QnAppServerFileCache::~QnAppServerFileCache() {
+}
 
 // -------------- Utility methods ----------------
 
 QString QnAppServerFileCache::getFullPath(const QString &filename) const {
+    auto connectionState = qnClientMessageProcessor->connectionState();
+    Q_ASSERT_X(connectionState != QnConnectionState::Disconnected || connectionState == QnConnectionState::Invalid,
+        Q_FUNC_INFO, "Method should be called only when we are know the target system. Current state is " + QnConnectionStateUtils::toString(connectionState).toUtf8());
+
+    /* Avoid empty folder name and collisions with our folders such as 'log'. */
+    QString systemName = L'_' + replaceNonFileNameCharacters(qnCommon->localSystemName(), L'_');
+
     QString path = QStandardPaths::writableLocation(QStandardPaths::DataLocation);
-    QUrl url = QnAppServerConnectionFactory::url();
-    return QDir::toNativeSeparators(QString(QLatin1String("%1/cache/%2_%3/%4/%5"))
+    return QDir::toNativeSeparators(QString(lit("%1/cache/%2/%3/%4"))
                                     .arg(path)
-                                    .arg(url.host(QUrl::FullyEncoded))
-                                    .arg(url.port())
+                                    .arg(systemName)
                                     .arg(m_folderName)
                                     .arg(filename)
                                     );
@@ -48,56 +85,64 @@ QString QnAppServerFileCache::folderName() const {
 
 void QnAppServerFileCache::clearLocalCache() {
     QString path = QStandardPaths::writableLocation(QStandardPaths::DataLocation);
-    QString dir = QDir::toNativeSeparators(QString(QLatin1String("%1/cache/")).arg(path));
+    QString dir = QDir::toNativeSeparators(QString(lit("%1/cache/")).arg(path));
     removeDir(dir);
+}
+
+bool QnAppServerFileCache::isConnectedToServer() const {
+    return QnAppServerConnectionFactory::getConnection2() != NULL;
+}
+
+qint64 QnAppServerFileCache::maximumFileSize() {
+    return ::maximumFileSize;
 }
 
 
 // -------------- File List loading methods -----
 
 void QnAppServerFileCache::getFileList() {
-    auto connection = QnAppServerConnectionFactory::getConnection2();
-    if (!connection) {
-        m_fileListHandle = -1;
+    if (!isConnectedToServer()) {
+        emit delayedFileListReceived(QStringList(), OperationResult::disconnected);
         return;
     }
 
-    m_fileListHandle = connection->getStoredFileManager()->listDirectory(
-                m_folderName,
-                this,
-                &QnAppServerFileCache::at_fileListReceived );
-}
-
-void QnAppServerFileCache::at_fileListReceived(int handle, ec2::ErrorCode errorCode, const QStringList& filenames) {
-    if (handle != m_fileListHandle)
-        return;
-
-    emit fileListReceived(filenames, errorCode == ec2::ErrorCode::ok);
+    auto connection = QnAppServerConnectionFactory::getConnection2();
+    connection->getStoredFileManager()->listDirectory(m_folderName, this, [this](int handle, ec2::ErrorCode errorCode, const QStringList& filenames) {
+        Q_UNUSED(handle);
+        bool ok = errorCode == ec2::ErrorCode::ok;
+        emit fileListReceived(filenames, ok ? OperationResult::ok : OperationResult::serverError);
+    } );
 }
 
 // -------------- Download File methods ----------
 
 void QnAppServerFileCache::downloadFile(const QString &filename) {
+    if (!isConnectedToServer()) {
+        emit delayedFileDownloaded(filename, OperationResult::disconnected);
+        return;
+    }
+
     if (filename.isEmpty()) {
-        emit delayedFileDownloaded(filename, false);
+        emit delayedFileDownloaded(filename, OperationResult::invalidOperation);
         return;
     }
 
     if (m_deleting.values().contains(filename)) {
-        emit delayedFileDownloaded(filename, false);
+        emit delayedFileDownloaded(filename, OperationResult::invalidOperation);
         return;
     }
 
     QFileInfo info(getFullPath(filename));
     if (info.exists()) {
-        emit delayedFileDownloaded(filename, true);
+        emit delayedFileDownloaded(filename, OperationResult::ok);
         return;
     }
 
     if (m_loading.values().contains(filename))
       return;
 
-    int handle = QnAppServerConnectionFactory::getConnection2()->getStoredFileManager()->getStoredFile(
+    auto connection = QnAppServerConnectionFactory::getConnection2();
+    int handle = connection->getStoredFileManager()->getStoredFile(
                 m_folderName + QLatin1Char('/') + filename,
                 this,
                 &QnAppServerFileCache::at_fileLoaded );
@@ -111,8 +156,13 @@ void QnAppServerFileCache::at_fileLoaded( int handle, ec2::ErrorCode errorCode, 
     QString filename = m_loading[handle];
     m_loading.remove(handle);
 
+    if (!isConnectedToServer()) {
+        emit fileDownloaded(filename, OperationResult::disconnected);
+        return;
+    }
+
     if (errorCode != ec2::ErrorCode::ok) {
-        emit fileDownloaded(filename, false);
+        emit fileDownloaded(filename, OperationResult::serverError);
         return;
     }
 
@@ -120,33 +170,44 @@ void QnAppServerFileCache::at_fileLoaded( int handle, ec2::ErrorCode errorCode, 
     QString filePath = getFullPath(filename);
     QFile file(filePath);
     if (!file.open(QIODevice::WriteOnly)) {
-        emit fileDownloaded(filename, false);
+        emit fileDownloaded(filename, OperationResult::fileSystemError);
         return;
     }
     QDataStream out(&file);
     out.writeRawData(data, data.size());
     file.close();
 
-    emit fileDownloaded(filename, true);
+    emit fileDownloaded(filename, OperationResult::ok);
 }
 
 // -------------- Uploading methods ----------------
 
 
 void QnAppServerFileCache::uploadFile(const QString &filename) {
+    if (!isConnectedToServer()) {
+        emit delayedFileUploaded(filename, OperationResult::disconnected);
+        return;
+    }
+
     if (m_uploading.values().contains(filename))
         return;
 
     QFile file(getFullPath(filename));
     if(!file.open(QIODevice::ReadOnly)) {
-        emit delayedFileUploaded(filename, false);
+        emit delayedFileUploaded(filename, OperationResult::fileSystemError);
+        return;
+    }
+
+    if(file.size() > ::maximumFileSize) {
+        emit delayedFileUploaded(filename, OperationResult::sizeLimitExceeded);
         return;
     }
 
     QByteArray data = file.readAll();
     file.close();
 
-    int handle = QnAppServerConnectionFactory::getConnection2()->getStoredFileManager()->addStoredFile(
+    auto connection = QnAppServerConnectionFactory::getConnection2();
+    int handle = connection->getStoredFileManager()->addStoredFile(
                 m_folderName + QLatin1Char('/') +filename,
                 data,
                 this,
@@ -161,17 +222,28 @@ void QnAppServerFileCache::at_fileUploaded( int handle, ec2::ErrorCode errorCode
 
     QString filename = m_uploading[handle];
     m_uploading.remove(handle);
+
+    if (!isConnectedToServer()) {
+        emit fileUploaded(filename, OperationResult::disconnected);
+        return;
+    }
+
     const bool ok = errorCode == ec2::ErrorCode::ok;
     if (!ok)
         QFile::remove(getFullPath(filename));
-    emit fileUploaded(filename, ok);
+    emit fileUploaded(filename, ok ? OperationResult::ok : OperationResult::serverError);
 }
 
 // -------------- Deleting methods ----------------
 
 void QnAppServerFileCache::deleteFile(const QString &filename) {
+    if (!isConnectedToServer()) {
+        emit delayedFileDeleted(filename, OperationResult::disconnected);
+        return;
+    }
+
     if (filename.isEmpty()) {
-        emit delayedFileDeleted(filename, false);
+        emit delayedFileDeleted(filename, OperationResult::invalidOperation);
         return;
     }
 
@@ -188,7 +260,8 @@ void QnAppServerFileCache::deleteFile(const QString &filename) {
     if (m_deleting.values().contains(filename))
       return;
 
-    int handle = QnAppServerConnectionFactory::getConnection2()->getStoredFileManager()->deleteStoredFile(
+    auto connection = QnAppServerConnectionFactory::getConnection2();
+    int handle = connection->getStoredFileManager()->deleteStoredFile(
                     m_folderName + QLatin1Char('/') +filename,
                     this,
                     &QnAppServerFileCache::at_fileDeleted );
@@ -202,8 +275,20 @@ void QnAppServerFileCache::at_fileDeleted( int handle, ec2::ErrorCode errorCode 
 
     QString filename = m_deleting[handle];
     m_deleting.remove(handle);
+
+    if (!isConnectedToServer()) {
+        emit fileDeleted(filename, OperationResult::disconnected);
+        return;
+    }
+    
     const bool ok = errorCode == ec2::ErrorCode::ok;
     if (ok)
         QFile::remove(getFullPath(filename));
-    emit fileDeleted(filename, ok);
+    emit fileDeleted(filename, ok ? OperationResult::ok : OperationResult::serverError);
+}
+
+void QnAppServerFileCache::clear() {
+    m_loading.clear();
+    m_uploading.clear();
+    m_deleting.clear();
 }
